array to function

+++ include/sensorsBME280.h
@@ -23,7 +23,7 @@ void readCCS811(void);
 void printSensorError(void);
 void printInfoSerial(void);
 void Charts_init(void);
-void array_temp_eCO2(int *array);
+void array_temp_eCO2(int16_t *array);
 
 extern float BMEtempC;
 extern int16_t eCO2;
